+++ libgeopmd/src/LevelZeroDevicePoolImp.hpp
@@ -103,7 +103,7 @@ namespace geopm
             void check_domain_exists(int size, const char *func, int line) const;
             std::pair<unsigned int, unsigned int> subdevice_device_conversion(unsigned int idx) const;
             mutable std::map<int, std::vector<uint64_t> > m_active_time_last; // Map from l0_domain to vector over gpu chips
-            mutable std::map<int, std::vector<int> > m_active_time_rollover; // Map from l0_domain to vector over gpu chips
+            mutable std::map<int, std::vector<uint64_t> > m_active_time_rollover; // Map from l0_domain to vector over gpu chips
     };
 }
 #endif
